# docs/swagger/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.CategoryResponse:
    properties:
      created_at:
        type: string
      created_by:
        type: string
      description:
        type: string
      id:
        type: string
      metadata:
        $ref: '#/definitions/types.Metadata'
      name:
        type: string
      slug:
        type: string
      status:
        $ref: '#/definitions/types.Status'
      updated_at:
        type: string
      updated_by:
        type: string
    type: object
...
  dto.CreateCategoryRequest:
    properties:
      description:
        maxLength: 2000
        type: string
      metadata:
        $ref: '#/definitions/types.Metadata'
      name:
        maxLength: 255
        minLength: 2
        type: string
      slug:
        maxLength: 100
        minLength: 3
        type: string
    required:
    - name
    - slug
    type: object
...
  dto.ListCategoriesResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/dto.CategoryResponse'
        type: array
      pagination:
        $ref: '#/definitions/types.PaginationResponse'
    type: object
...
  dto.UpdateCategoryRequest:
    properties:
      description:
        maxLength: 2000
        type: string
      name:
        maxLength: 255
        minLength: 2
        type: string
      slug:
        maxLength: 100
        minLength: 3
        type: string
    type: object
...
  ierr.ErrorDetail:
    properties:
      details:
        additionalProperties: {}
        type: object
      internal_error:
        type: string
      message:
        type: string
    type: object
  ierr.ErrorResponse:
    properties:
      error:
        $ref: '#/definitions/ierr.ErrorDetail'
      success:
        type: boolean
    type: object
...
  types.Metadata:
    additionalProperties:
      type: string
    type: object
  types.PaginationResponse:
    properties:
      limit:
        type: integer
      offset:
        type: integer
      total:
        type: integer
    type: object
...
  types.Status:
    enum:
    - published
    - draft
    - archived
    - deleted
    type: string
    x-enum-varnames:
    - StatusPublished
    - StatusDraft
    - StatusArchived
    - StatusDeleted
...
host: 5p9ubi66hh.execute-api.ap-south-1.amazonaws.com
info:
  contact:
    email: support@nashikdarshan.com
    name: API Support
  description: API for Nashik Darshan
  termsOfService: http://nashikdarshan.com/terms/
  title: Nashik Darshan API
  version: "1.0"
paths:
...
  /categories:
    get:
      consumes:
      - application/json
      description: Get a paginated list of categories with filtering and pagination
      parameters:
      - description: Limit
        in: query
        name: limit
        type: integer
      - description: Offset
        in: query
        name: offset
        type: integer
      - description: Status
        in: query
        name: status
        type: string
      - description: Sort field
        in: query
        name: sort
        type: string
      - description: Sort order (asc/desc)
        in: query
        name: order
        type: string
      - collectionFormat: csv
        description: Filter by slugs
        in: query
        items:
          type: string
        name: slug
        type: array
      - collectionFormat: csv
        description: Filter by names
        in: query
        items:
          type: string
        name: name
        type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ListCategoriesResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ierr.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ierr.ErrorResponse'
      summary: List categories
      tags:
      - Category
    post:
      consumes:
      - application/json
      description: Create a new category with the provided details
      parameters:
      - description: Create category request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateCategoryRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.CategoryResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ierr.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/ierr.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ierr.ErrorResponse'
      security:
      - Authorization: []
      - ApiKeyAuth: []
      summary: Create a new category
      tags:
      - Category
  /categories/{id}:
    delete:
      consumes:
      - application/json
      description: Soft delete a category
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ierr.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ierr.ErrorResponse'
      security:
      - Authorization: []
      - ApiKeyAuth: []
      summary: Delete a category
      tags:
      - Category
    get:
      consumes:
      - application/json
      description: Get a category by its ID
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.CategoryResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ierr.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ierr.ErrorResponse'
      summary: Get category by ID
      tags:
      - Category
    put:
      consumes:
      - application/json
      description: Update an existing category
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: string
      - description: Update category request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateCategoryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.CategoryResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ierr.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ierr.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/ierr.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ierr.ErrorResponse'
      security:
      - Authorization: []
      - ApiKeyAuth: []
      summary: Update a category
      tags:
      - Category
  /categories/slug/{slug}:
    get:
      consumes:
      - application/json
      description: Get a category by its slug
      parameters:
      - description: Category slug
        in: path
        name: slug
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.CategoryResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ierr.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ierr.ErrorResponse'
      summary: Get category by slug
      tags:
      - Category
...
schemes:
- https
securityDefinitions:
  ApiKeyAuth:
    description: API key for server-to-server or script access. Use X-API-Key header
      (alternative to Bearer).
    in: header
    name: X-API-Key
    type: apiKey
  Authorization:
    description: JWT Bearer token, e.g. `Bearer <token>`.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"